finish logic to handle divide by zero
When divide by zero , the operate function return undefined and the
evaluateAndDisplay function will handle the case so that I keep the operate
function remain to be pure function and evaluateAndDisplay function retian
to be side-effect function.

I also fix the bugs when implementing the logic. push the number before
call the operate function. call the treatMathError function but not
restart function .