generalize the function for button
The program receive user input , decide the input type , then perform action.
The action for keybroad and button are the same. So what I'm going to do
is to generalize the decide process. I create a function that handle button
and keybroad event and decide the input type , return an object. The program \
only perform action base on this object, but not specific to the button or keybroad
event statement so the code can be less repeated.